Java Music Application
This proof of concept application uses OOP practices, assisted with APIs to communicate with a database file using SQL queries to manipulate data in the database. All the resources needed to test and run this application is in this zip file. The IDE used to develop this software NetBeans assisted with Microsft Access for database administration practices.